+++ src/backend/marsha/settings.py
@@ -365,10 +365,11 @@ class Base(Configuration):
     EMAIL_FROM = values.Value("from@fun-mooc.fr")
 
     # REMINDERS SENT for scheduled webinars
-    REMINDER_1, REMINDER_2, REMINDER_3, REMINDER_ERROR = (
+    REMINDER_1, REMINDER_2, REMINDER_3, REMINDER_DATE_UPDATED, REMINDER_ERROR = (
         "REMINDER_1",
         "REMINDER_2",
         "REMINDER_3",
+        "REMINDER_DATE_UPDATED",
         "REMINDER_ERROR",
     )
     # keys for REMINDERS_STEP